+++ api/mojang/queue/queue.go
@@ -14,9 +14,10 @@ import (
 	"github.com/elyby/chrly/api/mojang"
 )
 
+var UuidsQueueIterationDelay = 2*time.Second + 500*time.Millisecond
+
 var usernamesToUuids = mojang.UsernamesToUuids
 var uuidToTextures = mojang.UuidToTextures
-var uuidsQueueIterationDelay = time.Second
 var forever = func() bool {
 	return true
 }
@@ -97,13 +98,13 @@ func (ctx *JobsQueue) GetTexturesForUsername(username string) chan *mojang.Signe
 
 func (ctx *JobsQueue) startQueue() {
 	go func() {
-		time.Sleep(uuidsQueueIterationDelay)
+		time.Sleep(UuidsQueueIterationDelay)
 		for forever() {
 			start := time.Now()
 			ctx.queueRound()
 			elapsed := time.Since(start)
 			ctx.Logger.RecordTimer("mojang_textures.usernames.round_time", elapsed)
-			time.Sleep(uuidsQueueIterationDelay)
+			time.Sleep(UuidsQueueIterationDelay)
 		}
 	}()
 }